Fixed dependency of get_physical_disk_number() on locale setting and removed code duplication.
